calculate gravity more intelligently; add minimum time for loop; don't use sg filter; add buffer for key speeds
Gravity is now calculated using the hammer travel in mm, and assumes
that the hammer will fall at 9.81ms^2. SG filter is commented out for
now, and no longer used in the for loop. The minimum time for the loop
is necessary to favourably shift the signal to noise ratio.