delete unused / irrelevant sketches
multi_note_simulation_fixed.ino was an attempt to fix the elapsed time
between iterations, so that dividing by elapsed was necessary in simulation
calculations. This was based on an older version of code that did not encapsulate
the simulation logic in a class.
multi_core_example was a simple multi core rpico exaple I committed early on
in this project when I thought I might use a multi-core setup.
multi_note_midi was an early file experimenting with getting midi working
with rpico.